create project

+++ 04.Methods-Exercise/08.FactorialDivision/Program.cs
@@ -6,21 +6,21 @@
         {
             int num1 = int.Parse(Console.ReadLine());
             int num2 = int.Parse(Console.ReadLine());
-            int factorial1 = FactorialNum1(num1);
-            int factorial2 = FactorialNum2(num2);
+            double factorial1 = FactorialNum1(num1);
+            double factorial2 = FactorialNum2(num2);
 
             PrintResult(factorial1, factorial2);
         }
 
-        private static void PrintResult(int factorial1, int factorial2)
+        private static void PrintResult(double factorial1, double factorial2)
         {
             double result = (double)factorial1 / factorial2;
             Console.WriteLine($"{result:f2}");
         }
 
-        private static int FactorialNum1(int num1)
+        private static double FactorialNum1(int num1)
         {
-            int factorial = 1;
+           double factorial = 1;
             for (int i = 0; i < num1; i++)
             {
                 factorial *= (num1 - i);
@@ -28,9 +28,9 @@
             return factorial;
         }
 
-        private static int FactorialNum2(int num2)
+        private static double FactorialNum2(int num2)
         {
-            int factorial = 1;
+            double factorial = 1;
             for (int i = 0; i < num2; i++)
             {
                 factorial *= (num2 - i);
